Resolve drone
