+++ openaev-front/src/admin/components/scenarios/scenario/ScenarioContext.ts
@@ -3,12 +3,8 @@ import { useState } from 'react';
 import { addInjectForScenario, bulkDeleteInjectsSimple, bulkUpdateInjectSimple, deleteInjectScenario, fetchScenarioInjects, updateInjectActivationForScenario, updateInjectForScenario } from '../../../../actions/Inject';
 import { bulkTestInjects } from '../../../../actions/inject_test/scenario-inject-test-actions';
 import { type InjectOutputType, type InjectStore } from '../../../../actions/injects/Inject';
-import {
-  importInjectsForScenario,
-  searchScenarioInjectsSimple,
-} from '../../../../actions/injects/inject-action';
 import { dryImportXlsForScenario, fetchScenario, fetchScenarioTeams, importXlsForScenario } from '../../../../actions/scenarios/scenario-actions';
-import { createInjectsForScenario } from '../../../../actions/scenarios/scenario-inject-actions';
+import { createInjectsForScenario, importInjectsForScenario, searchScenarioInjectsSimple } from '../../../../actions/scenarios/scenario-inject-actions';
 import { type Page } from '../../../../components/common/queryable/Page';
 import { type ImportTestSummary, type Inject, type InjectBulkProcessingInput, type InjectBulkUpdateInputs, type InjectInput, type InjectsImportInput, type InjectTestStatusOutput, type Scenario, type SearchPaginationInput } from '../../../../utils/api-types';
 import { useAppDispatch } from '../../../../utils/hooks';
