+++ openaev-front/src/admin/components/simulations/simulation/ExerciseContext.ts
@@ -15,11 +15,7 @@ import {
 } from '../../../../actions/Inject';
 import { bulkTestInjects } from '../../../../actions/inject_test/simulation-inject-test-actions';
 import { type InjectOutputType, type InjectStore } from '../../../../actions/injects/Inject';
-import {
-  importInjectsForSimulation,
-  searchExerciseInjectsSimple,
-} from '../../../../actions/injects/inject-action';
-import { createInjectsForSimulation } from '../../../../actions/simulations/simulation-inject-actions';
+import { createInjectsForSimulation, importInjectsForSimulation, searchExerciseInjectsSimple } from '../../../../actions/simulations/simulation-inject-actions';
 import { type Page } from '../../../../components/common/queryable/Page';
 import {
   type Exercise,
